sua api

+++ QuanLyPhongMayBackEnd/QuanLyPhongMayBackEnd/src/main/java/com/example/QuanLyPhongMayBackEnd/service/GhiChuMayTinhService.java
@@ -18,7 +18,10 @@ public class GhiChuMayTinhService {
     private boolean isUserLoggedIn(String token) {
         return taiKhoanService.checkUserLoginStatus(token).get("status").equals("success");
     }
-    public GhiChuMayTinh layGhiChuTheoMa(Long maGhiChuMT) {
+    public GhiChuMayTinh layGhiChuTheoMa(Long maGhiChuMT, String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         GhiChuMayTinh ghiChuMayTinh = null;
         Optional<GhiChuMayTinh> kq = ghiChuMayTinhRepository.findById(maGhiChuMT);
         try {
@@ -30,43 +33,70 @@ public class GhiChuMayTinhService {
     }
 
     @Transactional
-    public void xoa(Long maGhiChuMT) {
+    public void xoa(Long maGhiChuMT, String token) {
+        if (!isUserLoggedIn(token)) {
+            return ; // Token không hợp lệ
+        }
         ghiChuMayTinhRepository.deleteById(maGhiChuMT);
     }
 
     @Transactional
-    public void xoaTheoMaMay(Long maMay) {
+    public void xoaTheoMaMay(Long maMay, String token) {
+        if (!isUserLoggedIn(token)) {
+            return; // Token không hợp lệ
+        }
         List<GhiChuMayTinh> dsGhiChuMayTinh = ghiChuMayTinhRepository.findByMayTinh_MaMay(maMay);
         for (GhiChuMayTinh ghiChuMayTinh : dsGhiChuMayTinh) {
             ghiChuMayTinhRepository.delete(ghiChuMayTinh);
         }
     }
 
-    public GhiChuMayTinh luu(GhiChuMayTinh ghiChuMayTinh) {
+    public GhiChuMayTinh luu(GhiChuMayTinh ghiChuMayTinh, String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuMayTinhRepository.save(ghiChuMayTinh);
     }
 
-    public List<GhiChuMayTinh> layDSGhiChu() {
+    public List<GhiChuMayTinh> layDSGhiChu(String token) {
+        if (!isUserLoggedIn(token)) {
+            return null;
+        }
         return ghiChuMayTinhRepository.findAll();
     }
 
-    public GhiChuMayTinh capNhat(GhiChuMayTinh ghiChuMayTinh) {
+    public GhiChuMayTinh capNhat(GhiChuMayTinh ghiChuMayTinh,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuMayTinhRepository.save(ghiChuMayTinh);
     }
 
-    public List<GhiChuMayTinh> layDSGhiChuTheoNgaySua(Date ngaySua) {
+    public List<GhiChuMayTinh> layDSGhiChuTheoNgaySua(Date ngaySua,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuMayTinhRepository.findByNgaySua(ngaySua);
     }
 
-    public List<GhiChuMayTinh> layDSGhiChuTheoNgayBaoLoi(Date ngayBaoLoi) {
+    public List<GhiChuMayTinh> layDSGhiChuTheoNgayBaoLoi(Date ngayBaoLoi,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuMayTinhRepository.findByNgayBaoLoi(ngayBaoLoi);
     }
 
-    public List<GhiChuMayTinh> layDSGhiChuTheoMayTinh(Long maMay) {
+    public List<GhiChuMayTinh> layDSGhiChuTheoMayTinh(Long maMay,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuMayTinhRepository.findByMayTinh_MaMay(maMay);
     }
 
-    public GhiChuMayTinh layGhiChuGanNhatTheoMayTinh(Long maMay) {
+    public GhiChuMayTinh layGhiChuGanNhatTheoMayTinh(Long maMay,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         List<GhiChuMayTinh> dsGhiChuMayTinh = ghiChuMayTinhRepository.findByMayTinh_MaMayOrderByNgayBaoLoiDesc(maMay);
         return dsGhiChuMayTinh.isEmpty() ? null : dsGhiChuMayTinh.get(0);
     }
