+++ QuanLyPhongMayBackEnd/QuanLyPhongMayBackEnd/src/main/java/com/example/QuanLyPhongMayBackEnd/service/GhiChuPhongMayService.java
@@ -18,7 +18,10 @@ public class GhiChuPhongMayService {
     private boolean isUserLoggedIn(String token) {
         return taiKhoanService.checkUserLoginStatus(token).get("status").equals("success");
     }
-    public GhiChuPhongMay layGhiChuTheoMa(Long maGhiChu) {
+    public GhiChuPhongMay layGhiChuTheoMa(Long maGhiChu,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         GhiChuPhongMay ghiChuPhongMay = null;
         Optional<GhiChuPhongMay> kq = ghiChuPhongMayRepository.findById(maGhiChu);
         try {
@@ -30,35 +33,59 @@ public class GhiChuPhongMayService {
     }
 
     @Transactional
-    public void xoa(Long maGhiChu) {
+    public void xoa(Long maGhiChu, String token) {
+        if (!isUserLoggedIn(token)) {
+            return ; // Token không hợp lệ
+        }
         ghiChuPhongMayRepository.deleteById(maGhiChu);
     }
 
-    public GhiChuPhongMay luu(GhiChuPhongMay ghiChuPhongMay) {
+    public GhiChuPhongMay luu(GhiChuPhongMay ghiChuPhongMay,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuPhongMayRepository.save(ghiChuPhongMay);
     }
 
-    public List<GhiChuPhongMay> layDSGhiChu() {
+    public List<GhiChuPhongMay> layDSGhiChu(String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuPhongMayRepository.findAll();
     }
 
-    public GhiChuPhongMay capNhat(GhiChuPhongMay ghiChuPhongMay) {
+    public GhiChuPhongMay capNhat(GhiChuPhongMay ghiChuPhongMay, String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuPhongMayRepository.save(ghiChuPhongMay);
     }
 
-    public List<GhiChuPhongMay> layDSGhiChuTheoNgaySua(Date ngaySua) {
+    public List<GhiChuPhongMay> layDSGhiChuTheoNgaySua(Date ngaySua, String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuPhongMayRepository.findByNgaySua(ngaySua);
     }
 
-    public List<GhiChuPhongMay> layDSGhiChuTheoNgayBaoLoi(Date ngayBaoLoi) {
+    public List<GhiChuPhongMay> layDSGhiChuTheoNgayBaoLoi(Date ngayBaoLoi, String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuPhongMayRepository.findByNgayBaoLoi(ngayBaoLoi);
     }
 
-    public List<GhiChuPhongMay> layDSGhiChuTheoPhongMay(Long maPhong) {
+    public List<GhiChuPhongMay> layDSGhiChuTheoPhongMay(Long maPhong, String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return ghiChuPhongMayRepository.findByPhongMay_MaPhong(maPhong);
     }
 
-    public GhiChuPhongMay layGhiChuGanNhatTheoPhongMay(Long maPhong) {
+    public GhiChuPhongMay layGhiChuGanNhatTheoPhongMay(Long maPhong, String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         List<GhiChuPhongMay> dsGhiChu = ghiChuPhongMayRepository.findByPhongMay_MaPhongOrderByNgayBaoLoiDesc(maPhong);
         return dsGhiChu.isEmpty() ? null : dsGhiChu.get(0);
     }
