+++ QuanLyPhongMayBackEnd/QuanLyPhongMayBackEnd/src/main/java/com/example/QuanLyPhongMayBackEnd/service/MonHocService.java
@@ -20,7 +20,10 @@ public class MonHocService {
     private boolean isUserLoggedIn(String token) {
         return taiKhoanService.checkUserLoginStatus(token).get("status").equals("success");
     }
-    public MonHoc layMonHocTheoMa(Long maMon) {
+    public MonHoc layMonHocTheoMa(Long maMon,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         MonHoc monHoc = null;
         Optional<MonHoc> kq = monHocRepository.findById(maMon);
         try {
@@ -31,16 +34,25 @@ public class MonHocService {
         }
     }
 
-    public List<MonHoc> layDSMonHoc() {
+    public List<MonHoc> layDSMonHoc(String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return monHocRepository.findAll();
     }
 
     @Transactional
-    public void xoa(Long maMon) {
+    public void xoa(Long maMon,String token) {
+        if (!isUserLoggedIn(token)) {
+            return ; // Token không hợp lệ
+        }
         monHocRepository.deleteById(maMon);
     }
 
-    public MonHoc luu(MonHoc monHoc) {
+    public MonHoc luu(MonHoc monHoc,String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return monHocRepository.save(monHoc);
     }
 }
