+++ QuanLyPhongMayBackEnd/QuanLyPhongMayBackEnd/src/main/java/com/example/QuanLyPhongMayBackEnd/service/QuyenService.java
@@ -15,24 +15,40 @@ public class QuyenService {
 
     @Autowired
     private QuyenRepository quyenRepository;
-
+    @Autowired
+    private TaiKhoanService taiKhoanService;
+    private boolean isUserLoggedIn(String token) {
+        return taiKhoanService.checkUserLoginStatus(token).get("status").equals("success");
+    }
     // Lấy quyền theo mã
-    public Quyen layQuyenTheoMa(Long maQuyen) {
+    public Quyen layQuyenTheoMa(Long maQuyen, String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return null; // Bạn có thể thêm logic tại đây
     }
 
     // Lấy danh sách quyền
-    public List<Quyen> layDSQuyen() {
+    public List<Quyen> layDSQuyen(String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return quyenRepository.findAll();
     }
 
     // Xóa quyền theo mã
-    public void xoa(Long maQuyen) {
+    public void xoa(Long maQuyen, String token) {
+        if (!isUserLoggedIn(token)) {
+            return ; // Token không hợp lệ
+        }
         quyenRepository.deleteById(maQuyen);
     }
 
     // Lưu quyền
-    public Quyen luu(Quyen quyen) {
+    public Quyen luu(Quyen quyen , String token) {
+        if (!isUserLoggedIn(token)) {
+            return null; // Token không hợp lệ
+        }
         return quyenRepository.save(quyen);
     }
 }
